fix sphinx demo

+++ accessory/model/LLM/llama_ens.py
@@ -396,8 +396,8 @@ class Transformer(nn.Module):
 
             local_clip_image_feats = self.clip_encode_image(local_image)
             local_convnext_image_feats = self.openclip_convnext_xxl(
-                F.interpolate(local_image.half(), size=(256, 256))
-            ).to(local_image)
+                F.interpolate(local_image.half(), size=(256, 256)).to(local_image)
+            )
             assert local_convnext_image_feats.size()[1:] == (3072, 8, 8)
             local_convnext_image_feats = local_convnext_image_feats.repeat_interleave(
                 2, dim=-1
@@ -436,8 +436,8 @@ class Transformer(nn.Module):
             dist.all_gather_into_tensor(ens_image_feats, local_ens_image_feats,
                                         group=fs_init.get_model_parallel_group())
 
-            ens_image_feats = ens_image_feats.float()[:image_bs]
-            image_feats = image_feats.float()[:image_bs]
+            ens_image_feats = ens_image_feats[:image_bs]
+            image_feats = image_feats[:image_bs]
 
         image_feats = self.qformer_proj(image_feats)
         ens_image_feats = self.visual_proj(ens_image_feats)
